+++ src/com/codecool/api/components/Hardware.java
@@ -6,7 +6,7 @@ import com.codecool.api.enums.Stuff;
 public abstract class Hardware extends Components {
     private Stuff madeBy;
     
-    public Hardware(String name, String producer, int value, AreaToUse qualified, Stuff madeBy) {
+    public Hardware(String name, String producer, double value, AreaToUse qualified, Stuff madeBy) {
         super(name, producer, value, qualified);
         this.madeBy = madeBy;
     }
