+++ src/com/codecool/api/components/KnobsAndPulls.java
@@ -10,7 +10,7 @@ public abstract class KnobsAndPulls extends Hardware {
     private Style style;
     private Finish color;
     
-    KnobsAndPulls(String name, String producer, int value, AreaToUse qualified, Stuff madeBy, Style style, Finish color) {
+    KnobsAndPulls(String name, String producer, double value, AreaToUse qualified, Stuff madeBy, Style style, Finish color) {
         super(name, producer, value, qualified, madeBy);
         this.style = style;
         this.color = color;
