+++ src/com/codecool/api/components/Lumber.java
@@ -6,7 +6,7 @@ import com.codecool.api.enums.Load;
 public class Lumber extends Wood {
     private String species;
     
-    Lumber(String name, String producer, Load load, int value, AreaToUse qualified, int length, int width, int thickness) {
+    Lumber(String name, String producer, Load load, double value, AreaToUse qualified, int length, int width, int thickness) {
         super(name, producer, load, value, qualified, length, width, thickness);
         this.species = species;
     }
