+++ src/com/codecool/api/components/MDF.java
@@ -5,7 +5,7 @@ import com.codecool.api.enums.Load;
 
 public class MDF extends MenMadeSheet {
     
-    MDF(String name, String producer, Load LOW, int value, AreaToUse TRAFICM, int length, int width, int thickness) {
+    MDF(String name, String producer, Load LOW, double value, AreaToUse TRAFICM, int length, int width, int thickness) {
         super(name, producer, LOW, value, TRAFICM, length, width, thickness);
     }
