+++ src/com/codecool/api/components/MenMadeSheet.java
@@ -5,7 +5,7 @@ import com.codecool.api.enums.Load;
 
 public abstract class MenMadeSheet extends Wood {
     
-    MenMadeSheet(String name, String producer, Load load, int value, AreaToUse qualified, int length, int width, int thickness) {
+    MenMadeSheet(String name, String producer, Load load, double value, AreaToUse qualified, int length, int width, int thickness) {
         super(name, producer, load, value, qualified, length, width, thickness);
     }
 }
