+++ src/com/codecool/api/components/PlyWood.java
@@ -4,7 +4,7 @@ import com.codecool.api.enums.AreaToUse;
 import com.codecool.api.enums.Load;
 
 public class PlyWood extends MenMadeSheet {
-    PlyWood(String name, String producer, Load MEDIUM, int value, AreaToUse TRAFICM, int length, int width, int thickness) {
+    PlyWood(String name, String producer, Load MEDIUM, double value, AreaToUse TRAFICM, int length, int width, int thickness) {
         super(name, producer, MEDIUM, value, TRAFICM, length, width, thickness);
     }
